Changed `gethostbyaddr` to ip range
Callbacks from WorldPay may come from any IP in the range 195.35.90.0 to 195.35.91.255. For many of these, `gethostbyaddr` returns the IP address, not a worldpay.com domain name (e.g. the Hostname value for 195.35.90.0 is 195.35.90.0 - https://whatismyipaddress.com/ip/195.35.90.0). This means that WorldPay is accepting the payment, performing the callback and failing! Orders in Magento are being listed as 'pending_payment' instead of changing to 'complete'.

This patch replaces `gethostbyaddr` with a conditional check of the IPs.